refactor: shared > ui > index.tsx에 작성되어 있던 ui 컴포넌트들 모두 역할별로 분리
