+++ src/shared/ui/Card/Card.tsx
@@ -1,0 +1,7 @@
+import { forwardRef } from "react"
+
+// 카드 컴포넌트
+export const Card = forwardRef(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
+Card.displayName = "Card"
