+++ src/shared/ui/Dialog/Dialog.tsx
@@ -1,0 +1,3 @@
+import * as DialogPrimitive from "@radix-ui/react-dialog"
+
+export const Dialog = DialogPrimitive.Root
