+++ src/shared/ui/Table/Table.tsx
@@ -1,0 +1,8 @@
+import { forwardRef } from "react"
+
+export const Table = forwardRef(({ className, ...props }, ref) => (
+  <div className="w-full overflow-auto">
+    <table ref={ref} className={`table-fixed w-full caption-bottom text-sm ${className}`} {...props} />
+  </div>
+))
+Table.displayName = "Table"
