+++ src/shared/ui/Table/TableRow.tsx
@@ -1,0 +1,10 @@
+import { forwardRef } from "react"
+
+export const TableRow = forwardRef(({ className, ...props }, ref) => (
+  <tr
+    ref={ref}
+    className={`border-b transition-colors hover:bg-muted/50 data-[state=selected]:bg-muted h-14 ${className}`}
+    {...props}
+  />
+))
+TableRow.displayName = "TableRow"
