+++ src/shared/ui/index.ts
@@ -1,0 +1,7 @@
+export { Button } from "./Button"
+export { Card, CardContent, CardHeader, CardTitle } from "./Card"
+export { Dialog, DialogContent, DialogHeader, DialogOverlay, DialogPortal, DialogTitle, DialogTrigger } from "./Dialog"
+export { Input } from "./Input"
+export { Select, SelectContent, SelectGroup, SelectItem, SelectTrigger, SelectValue } from "./Select"
+export { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "./Table"
+export { Textarea } from "./Textarea"
